refactor(SD-ARCH-HOTSPOT-STAGE-WORKER-001): relocate S11/S15/S17 post-stage hooks to stage-handlers/ registry
FR-1: formatDuration -> lib/eva/util.js (verbatim; sibling copies deliberately
not deduped — they format differently, out of scope).
FR-2: seven _postStageHook_S{11,15,17}_* bodies relocated VERBATIM to
lib/eva/stage-handlers/{s11,s15,s17}.js (mechanical transforms only:
this._supabase/this._logger -> ctx.*, dynamic imports re-pathed one level up).
Load-bearing hook ORDER and its rationale comments relocated into each
module's execute().
FR-3: lib/eva/stage-handlers/registry.js — externalized registry (pattern
proven by lib/sweep/pass-registry.cjs) with STAGE_HANDLER_REGISTRY=off
kill-switch. Both dispatch paths call the SAME handler modules; the switch
toggles mechanism, never behavior. Fail-soft posture matches
_runPostStageHooks verbatim. S19 stays in-worker (deep _runS19Bridge
coupling — PRD conditional-relocation clause).

Worker: 4,893 -> 4,260 lines. Behavior preservation proven by the frozen
pre-refactor IO snapshot equivalence tests (committed pre-refactor in
9a17444bf55) — see follow-up test commit.

LOC justification (>400): pure relocation refactor; net-new logic is the
59-line registry + 24-line util + ctx helper. Every relocated line is
byte-tracked by the snapshot equivalence suite.


Fleet-Worker: Bravo
