+++ lib/eva/stage-execution-worker.js
@@ -49,6 +49,16 @@ import { toValidTransitionType } from './transition-type.js';
 // SD-LEO-INFRA-HARDEN-S19-S20-001 (FR-1): the single pure bridge-outcome classifier + S19
 // hold/advance decision shared by every S19 gate consumer (closes RCA 7610876f schism).
 import { classifyBridgeOutcome, shouldHoldAtS19, S19_BRIDGE_OUTCOME } from './bridge/s19-advance-decision.js';
+// SD-ARCH-HOTSPOT-STAGE-WORKER-001: post-stage hook bodies for S11/S15/S17 relocated to
+// lib/eva/stage-handlers/ (FR-2); dispatch via the externalized registry (FR-3). Both the
+// registry path and the kill-switch (STAGE_HANDLER_REGISTRY=off) legacy Map path call the
+// SAME handler modules — single implementation, dispatch mechanism is the only toggle.
+import * as stageHandlerRegistry from './stage-handlers/registry.js';
+import * as stageHandlersS11 from './stage-handlers/s11.js';
+import * as stageHandlersS15 from './stage-handlers/s15.js';
+import * as stageHandlersS17 from './stage-handlers/s17.js';
+// SD-ARCH-HOTSPOT-STAGE-WORKER-001 (FR-1): formatDuration relocated to shared util.
+import { formatDuration } from './util.js';
 // SD-LEO-INFRA-VISION-GROUNDED-ACCEPTANCE-001 (FR-2): vision-grounded acceptance gate (pure decision).
 import { shouldHoldForVisionAcceptance, isVisionAcceptanceGateEnabled, isVisionAcceptanceStrict } from './bridge/vision-acceptance-gate.js';
 import { shouldHoldForVisionDrift, isVisionDriftGateEnabled, isVisionDriftStrict, DRIFT_HOLD_CAUSE } from './bridge/vision-drift-gate.js';
@@ -104,22 +114,6 @@ function getOperatingMode(stage) {
     if (stage >= range.start && stage <= range.end) return mode;
   }
   return 'UNKNOWN';
-}
-
-/**
- * Format milliseconds into human-readable duration (e.g., "2m 30s").
- * @param {number} ms
- * @returns {string}
- */
-function formatDuration(ms) {
-  const seconds = Math.floor(ms / 1000);
-  if (seconds < 60) return `${seconds}s`;
-  const minutes = Math.floor(seconds / 60);
-  const remainingSec = seconds % 60;
-  if (minutes < 60) return `${minutes}m ${remainingSec}s`;
-  const hours = Math.floor(minutes / 60);
-  const remainingMin = minutes % 60;
-  return `${hours}h ${remainingMin}m`;
 }
 
 /**
@@ -211,28 +205,15 @@ export class StageExecutionWorker {
      * @type {Map<number, (ventureId: string) => Promise<void>>}
      */
     this._postStageHookRegistry = new Map([
-      // SD-LEO-FEAT-GVOS-ACTIVATION-REMEDIATION-001/FR-4: chain S11 hooks.
-      // GvosProfile runs AFTER logo so classifier errors do not block logo rendering.
-      // SD-LEO-FEAT-VENTURE-BRAND-GROUNDED-001/FR-1: NamePromotion runs first so the
-      // promoted brand name is in place for the logo fallback prompt and downstream
-      // consumers. Each hook is independently non-throwing, so ordering is failure-safe.
-      [11, async (ventureId) => {
-        await this._postStageHook_S11_NamePromotion(ventureId);
-        await this._postStageHook_S11_LogoGeneration(ventureId);
-        await this._postStageHook_S11_GvosProfile(ventureId);
-      }],
-      [15, (ventureId) => this._postStageHook_S15_StitchProvision(ventureId)],
-      // SD-LEO-REFAC-RETIRE-LEGACY-STAGE-001 (GVOS FR-6): freeze the design snapshot at the S17
-      // transition (idempotent, non-throwing) BEFORE doc-gen, so the lock fires regardless of the
-      // S17 strategy gate and for RPC-auto-advanced ventures the browser-side locker never reached.
-      [17, async (ventureId) => {
-        await this._postStageHook_S17_GvosLock(ventureId);
-        await this._postStageHook_S17_DocGen(ventureId);
-        // SD-LEO-INFRA-RELIABLE-S19-BUILD-001 / FR-1: auto-draft a draft_seed L2 vision so
-        // leo_bridge ventures have SOMETHING for the S19 gate to escalate on (and a human to
-        // approve) instead of silently producing 0 SDs. Non-throwing.
-        await this._postStageHook_S17_SeedDraftVision(ventureId);
-      }],
+      // SD-ARCH-HOTSPOT-STAGE-WORKER-001: S11/S15/S17 entries call the relocated
+      // handler modules (lib/eva/stage-handlers/) — the load-bearing hook ORDER and
+      // its rationale live in each module's execute(). This Map is the kill-switch
+      // (STAGE_HANDLER_REGISTRY=off) dispatch path; the enabled path routes through
+      // stage-handlers/registry.js runStageHandler in _runPostStageHooks. S19 stays
+      // in-worker (deep _runS19Bridge coupling — PRD conditional-relocation clause).
+      [11, (ventureId) => stageHandlersS11.execute(this._stageHandlerCtx(ventureId))],
+      [15, (ventureId) => stageHandlersS15.execute(this._stageHandlerCtx(ventureId))],
+      [17, (ventureId) => stageHandlersS17.execute(this._stageHandlerCtx(ventureId))],
       [19, (ventureId) => this._postStageHook_S19_Bridge(ventureId)],
     ]);
   }
@@ -2999,7 +2980,25 @@ export class StageExecutionWorker {
    * @param {string} ventureId
    * @param {number} completedStage - The stage that just finished
    */
+  /** ctx surface the relocated stage handlers consume (FR-2 enumeration). */
+  _stageHandlerCtx(ventureId) {
+    return {
+      supabase: this._supabase,
+      logger: this._logger,
+      ventureId,
+      ensureS17StrategySelected: (vId) => this._ensureS17StrategySelected(vId),
+    };
+  }
+
   async _runPostStageHooks(ventureId, completedStage) {
+    // SD-ARCH-HOTSPOT-STAGE-WORKER-001 (FR-3): externalized-registry dispatch first;
+    // the internal Map remains the kill-switch path AND the home of S19 (not in the
+    // external registry). runStageHandler is fail-soft with this method's exact
+    // posture, so behavior is identical on both paths.
+    if (stageHandlerRegistry.isRegistryEnabled()) {
+      const handled = await stageHandlerRegistry.runStageHandler(completedStage, this._stageHandlerCtx(ventureId));
+      if (handled) return;
+    }
     const hook = this._postStageHookRegistry.get(completedStage);
     if (!hook) return;
     try {
@@ -3010,638 +3009,6 @@ export class StageExecutionWorker {
     }
   }
 
-  /**
-   * S17 post-stage hook: Generate EVA vision + architecture plan documents.
-   * @param {string} ventureId
-   */
-  /**
-   * S11 post-stage hook: Generate logo image from logoSpec via Imagen 3.
-   * SD-EVA-FEAT-LOGO-IMAGEN-PIPELINE-001
-   *
-   * Viability gate: only renders for ventures past S7 (strategy formulation).
-   * Idempotent: skips if logo already exists in venture_artifacts.
-   */
-  async _postStageHook_S11_LogoGeneration(ventureId) {
-    const logger = this._logger;
-
-    // Viability gate: check venture has passed S7
-    const { data: stages } = await this._supabase
-      .from('venture_stage_work')
-      .select('lifecycle_stage')
-      .eq('venture_id', ventureId)
-      .gte('lifecycle_stage', 7)
-      .limit(1);
-    if (!stages || stages.length === 0) {
-      logger.log('[S11-Logo] Skipping — venture has not passed S7');
-      return;
-    }
-
-    // Idempotency: check if logo already exists
-    const { data: existing } = await this._supabase
-      .from('venture_artifacts')
-      .select('id')
-      .eq('venture_id', ventureId)
-      .eq('artifact_type', 'identity_logo_image')
-      .limit(1);
-    if (existing && existing.length > 0) {
-      logger.log('[S11-Logo] Skipping — logo already exists');
-      return;
-    }
-
-    // SD-LEO-FEAT-VENTURE-BRAND-GROUNDED-001/FR-3: read logoSpec from the canonical
-    // identity_brand_name artifact, not venture_stage_work.stage_data. The S11 analysis
-    // writer (stage-templates/analysis-steps/stage-11-visual-identity.js) persists logoSpec
-    // inside identity_brand_name.artifact_data.logoSpec — it is NOT reliably mirrored into
-    // stage_data, which silently skipped logo generation for ventures that had a valid spec.
-    // The `artifact_data` JSONB column is the writeArtifact mapping (artifact-persistence-service.js).
-    let logoSpec = null;
-    const { data: brandArtifact } = await this._supabase
-      .from('venture_artifacts')
-      .select('artifact_data')
-      .eq('venture_id', ventureId)
-      .eq('artifact_type', 'identity_brand_name')
-      .order('updated_at', { ascending: false })
-      .limit(1)
-      .maybeSingle();
-    logoSpec = brandArtifact?.artifact_data?.logoSpec || null;
-
-    // SD-REFILL-007PVF5E: removed the dead legacy venture_stage_work.stage_data fallback — that
-    // column does not exist live (the read threw), and logoSpec is canonically sourced from the
-    // identity_brand_name artifact above (SD-LEO-FEAT-VENTURE-BRAND-GROUNDED-001/FR-3), so the
-    // fallback could never have recovered a spec the canonical read missed.
-    if (!logoSpec) {
-      logger.log('[S11-Logo] No logoSpec in identity_brand_name artifact — skipping');
-      return;
-    }
-
-    // Get venture name for fallback prompt
-    const { data: venture } = await this._supabase
-      .from('ventures')
-      .select('name')
-      .eq('id', ventureId)
-      .single();
-
-    // Render logo via Imagen 3
-    const { renderLogo } = await import('./bridge/imagen-logo-renderer.js');
-    const result = await renderLogo(logoSpec, {
-      ventureName: venture?.name || 'Venture',
-      maxRetries: 2,
-      logger,
-    });
-
-    if (!result) {
-      logger.warn('[S11-Logo] Logo generation failed — venture will use text fallback');
-      return;
-    }
-
-    // Upload to Supabase Storage
-    const storagePath = `logos/${ventureId}/logo.png`;
-    const { error: uploadErr } = await this._supabase.storage
-      .from('venture-logos')
-      .upload(storagePath, result.buffer, {
-        contentType: result.mimeType,
-        upsert: true,
-      });
-
-    if (uploadErr) {
-      // Bucket may not exist yet — try to create it
-      if (uploadErr.message?.includes('not found') || uploadErr.statusCode === 404) {
-        logger.log('[S11-Logo] Creating venture-logos bucket');
-        await this._supabase.storage.createBucket('venture-logos', { public: true });
-        const { error: retryErr } = await this._supabase.storage
-          .from('venture-logos')
-          .upload(storagePath, result.buffer, { contentType: result.mimeType, upsert: true });
-        if (retryErr) {
-          logger.warn('[S11-Logo] Upload failed after bucket creation:', retryErr.message);
-          return;
-        }
-      } else {
-        logger.warn('[S11-Logo] Upload failed:', uploadErr.message);
-        return;
-      }
-    }
-
-    // Get public URL
-    const { data: urlData } = this._supabase.storage
-      .from('venture-logos')
-      .getPublicUrl(storagePath);
-
-    const logoUrl = urlData?.publicUrl;
-
-    // Write logo artifact
-    const { writeArtifact } = await import('./artifact-persistence-service.js');
-    await writeArtifact(this._supabase, {
-      ventureId,
-      lifecycleStage: 11,
-      artifactType: 'identity_logo_image',
-      title: 'Logo Image (Imagen 3)',
-      artifactData: { logoUrl, logoSpec, storagePath, generatedAt: new Date().toISOString() },
-      metadata: { source: 'post-stage-11-hook', renderer: 'imagen-3' },
-      source: 'post-stage-11-logo-hook',
-    });
-
-    logger.log('[S11-Logo] Logo generated and stored', { logoUrl, storagePath });
-  }
-
-  /**
-   * S11 post-stage hook: Write venture_gvos_profile row using the rule-only classifier.
-   *
-   * SD-LEO-FEAT-GVOS-ACTIVATION-REMEDIATION-001 / FR-4
-   *
-   * Guarantees: zero @anthropic-ai/sdk invocations (rule_only mode). Idempotent
-   * via UPSERT on venture_id. If venture lacks the inputs the rule classifier
-   * needs (business_model_class, industry_tags, audience_tags), falls through
-   * to emergency_default with chairman_warning audit log.
-   *
-   * GvosS17Sections gates rendering on useVentureArchetype(ventureId).data !== null;
-   * this hook is the producer that makes the activation chain visible at S17.
-   */
-  async _postStageHook_S11_GvosProfile(ventureId) {
-    const logger = this._logger;
-
-    try {
-      // Idempotency: skip if profile row already exists
-      const { data: existing, error: existErr } = await this._supabase
-        .from('venture_gvos_profile')
-        .select('id, archetype_id')
-        .eq('venture_id', ventureId)
-        .maybeSingle();
-
-      if (existErr && existErr.code !== 'PGRST116') {
-        logger.warn('[S11-GvosProfile] existence check failed:', existErr.message);
-        return;
-      }
-
-      if (existing && existing.archetype_id) {
-        logger.log('[S11-GvosProfile] Skipping — profile already exists for venture', ventureId);
-        return;
-      }
-
-      // Load venture inputs for the rule-based classifier
-      const { data: venture, error: ventureErr } = await this._supabase
-        .from('ventures')
-        .select('id, business_model_class, industry, vertical_category, tags, target_market')
-        .eq('id', ventureId)
-        .maybeSingle();
-
-      if (ventureErr || !venture) {
-        logger.warn('[S11-GvosProfile] venture load failed:', ventureErr?.message || 'venture not found');
-        return;
-      }
-
-      // Call rule-only classifier (JS port of EHG/src/lib/gvos/auto-classifier.ts mode='rule_only').
-      // Cross-repo parity contract documented in lib/gvos/rule-classifier.js.
-      const { classifyArchetypeRuleOnly } = await import('../../lib/gvos/rule-classifier.js');
-      const { buildClassifierInputFromVenture } = await import('../../lib/gvos/venture-classifier-inputs.js');
-      const result = await classifyArchetypeRuleOnly(
-        buildClassifierInputFromVenture(venture),
-        this._supabase,
-      );
-
-      // archetype_selection_method enum (schema CHECK):
-      // rule_based | llm_fallback | rule_fallback_below_threshold | emergency_default | chairman_override.
-      // Pass through the classifier's method verbatim — no normalization needed.
-      const profileRow = {
-        venture_id: ventureId,
-        version: 1,
-        archetype_id: result.archetypeId || null,
-        archetype_selection_method: result.method,
-        archetype_selection_confidence: result.confidence,
-        archetype_selection_rationale: result.rationale,
-        business_model_class: venture.business_model_class || null,
-      };
-
-      // onConflict MUST match the table's actual unique constraint (venture_id, version).
-      // Using 'venture_id' alone threw Postgres 42P10 and wrote 0 rows, which — combined
-      // with the swallowed error below — left every new venture without a GVOS profile and
-      // blanked Stage 17 (RCA 2026-05-22).
-      const { error: upsertErr } = await this._supabase
-        .from('venture_gvos_profile')
-        .upsert(profileRow, { onConflict: 'venture_id,version' });
-
-      if (upsertErr) {
-        // Surface loudly: a silently-swallowed write here previously hid a 100%-reproducible
-        // failure that blanked Stage 17 for all new ventures.
-        logger.error('[S11-GvosProfile] profile upsert FAILED:', upsertErr.message);
-        return;
-      }
-
-      // Low-confidence path → chairman_warning audit row (visible for manual override
-      // via SQL runbook until the S11 chairman-override UI ships in a follow-up SD).
-      if (result.confidence < 0.85) {
-        // SD-REFILL-007PVF5E: conform to the live audit_log schema (mirrors lib/eva/forward-gate.js):
-        // event_subtype/venture_id/details do not exist as columns; entity_type/entity_id are NOT NULL.
-        // Fold the former phantom fields into metadata and address the row to the venture entity.
-        const { error: warnErr } = await this._supabase.from('audit_log').insert({
-          event_type: 'chairman_warning',
-          entity_type: 'venture',
-          entity_id: ventureId,
-          severity: 'warning',
-          metadata: {
-            event_subtype: 'gvos_low_confidence_archetype',
-            venture_id: ventureId,
-            archetype_prompt_token: result.archetypePromptToken,
-            confidence: result.confidence,
-            rationale: result.rationale,
-            sd: 'SD-LEO-FEAT-GVOS-ACTIVATION-REMEDIATION-001',
-          },
-        });
-        if (warnErr && !/audit_log/i.test(warnErr.message || '')) {
-          // audit_log may not exist in all environments; non-fatal
-          logger.warn('[S11-GvosProfile] chairman_warning insert failed:', warnErr.message);
-        }
-      }
-
-      logger.log('[S11-GvosProfile] profile written', {
-        ventureId,
-        archetype: result.archetypePromptToken,
-        method: result.method,
-        confidence: result.confidence,
-      });
-    } catch (err) {
-      // Never throw — S11-LogoGeneration must remain the chairman-visible path.
-      logger.warn('[S11-GvosProfile] hook errored (non-fatal):', err.message);
-    }
-  }
-
-  /**
-   * S11 post-stage hook: promote the chosen brand name into ventures.name.
-   *
-   * SD-LEO-FEAT-VENTURE-BRAND-GROUNDED-001 / FR-1
-   *
-   * Source of truth: identity_brand_name.artifact_data.decision.selectedName.
-   *
-   * PROVENANCE GUARD (no schema change).
-   * LEAD risk evidence: ventures.name has a large blast radius (it threads into
-   * repo names, prompts, docs, marketing copy and chairman-facing surfaces) AND
-   * the system has NO placeholder-name convention — at creation the name is
-   * auto-derived (discovery synthesis) and stored ONLY in ventures.name, with no
-   * separate original-name column, no name_source column, and no name-change audit
-   * row (verified live 2026-05-22). A naive "looks like a placeholder" string check
-   * is therefore unsafe, so this hook self-stamps its own provenance instead:
-   *
-   *   metadata.brand_name_promotion = { promoted_name, prior_name, promoted_at, sd,
-   *                                     source_selected_name }
-   *
-   * Decision table:
-   *   - No marker (first run): the auto-derived creation name has never been promoted.
-   *     Promote selectedName -> ventures.name and stamp the marker (records prior_name
-   *     for audit). This is the single intended promotion window; it runs early and
-   *     automatically in the same pipeline pass that produced the brand, before any
-   *     chairman name-edit surface exists.
-   *   - Marker present AND ventures.name === marker.promoted_name: the current name is
-   *     still exactly the value WE last wrote, so it is safe. Refresh to a changed
-   *     selectedName if needed (no write when already equal -> idempotent).
-   *   - Marker present AND ventures.name !== marker.promoted_name: a human/chairman
-   *     deliberately changed the name after our promotion. STOP permanently and never
-   *     overwrite. This is the chairman-edit detector that satisfies the FR-1 invariant.
-   *
-   * Residual risk (documented): a human edit made BETWEEN creation and the very first
-   * S11 promotion cannot be distinguished from the auto-derived creation name because
-   * no creation-name baseline is persisted anywhere. This is mitigated by promotion
-   * running early/automatically in-pipeline and by the no-op short-circuit when
-   * selectedName already equals the current name. Closing it fully would require a
-   * schema change (a creation-name/name_source column), which is out of scope here.
-   *
-   * Non-throwing / non-blocking: mirrors the GvosProfile hook posture — any error is
-   * logged and swallowed so it never blocks stage advancement.
-   *
-   * @param {string} ventureId
-   */
-  async _postStageHook_S11_NamePromotion(ventureId) {
-    const logger = this._logger;
-    const SD_TAG = 'SD-LEO-FEAT-VENTURE-BRAND-GROUNDED-001';
-
-    try {
-      // Read the chosen name from the canonical identity_brand_name artifact.
-      const { data: brandArtifact } = await this._supabase
-        .from('venture_artifacts')
-        .select('artifact_data')
-        .eq('venture_id', ventureId)
-        .eq('artifact_type', 'identity_brand_name')
-        .order('updated_at', { ascending: false })
-        .limit(1)
-        .maybeSingle();
-
-      const selectedName = brandArtifact?.artifact_data?.decision?.selectedName;
-      if (!selectedName || typeof selectedName !== 'string' || !selectedName.trim()) {
-        logger.log('[S11-NamePromotion] No decision.selectedName in identity_brand_name artifact — skipping');
-        return;
-      }
-      const chosen = selectedName.trim();
-
-      // Load current venture name + provenance marker.
-      const { data: venture, error: ventureErr } = await this._supabase
-        .from('ventures')
-        .select('id, name, metadata')
-        .eq('id', ventureId)
-        .maybeSingle();
-
-      if (ventureErr || !venture) {
-        logger.warn('[S11-NamePromotion] venture load failed:', ventureErr?.message || 'venture not found');
-        return;
-      }
-
-      const metadata = venture.metadata || {};
-      const marker = metadata.brand_name_promotion || null;
-      const currentName = venture.name || null;
-
-      if (marker) {
-        // Provenance guard: only the value we ourselves last promoted is safe to touch.
-        if (currentName !== marker.promoted_name) {
-          logger.log(
-            '[S11-NamePromotion] Skipping — venture name was changed after promotion ' +
-            `(deliberate/chairman edit detected). current=${JSON.stringify(currentName)} ` +
-            `last_promoted=${JSON.stringify(marker.promoted_name)}`
-          );
-          return;
-        }
-        if (currentName === chosen) {
-          logger.log('[S11-NamePromotion] Skipping — name already promoted to selected value (idempotent)');
-          return;
-        }
-        // Safe refresh: name still equals our last promotion but selectedName changed.
-      }
-
-      // Promote (first run, or safe refresh).
-      const newMarker = {
-        promoted_name: chosen,
-        prior_name: currentName,
-        promoted_at: new Date().toISOString(),
-        sd: SD_TAG,
-        source_selected_name: chosen,
-      };
-
-      const { error: updateErr } = await this._supabase
-        .from('ventures')
-        .update({ name: chosen, metadata: { ...metadata, brand_name_promotion: newMarker } })
-        .eq('id', ventureId);
-
-      if (updateErr) {
-        logger.warn('[S11-NamePromotion] ventures.name promotion update failed:', updateErr.message);
-        return;
-      }
-
-      logger.log('[S11-NamePromotion] Promoted brand name', {
-        ventureId,
-        from: currentName,
-        to: chosen,
-      });
-    } catch (err) {
-      // Never throw — name promotion must never block stage advancement.
-      logger.warn('[S11-NamePromotion] hook errored (non-fatal):', err.message);
-    }
-  }
-
-  /**
-   * S15 post-stage hook: Store wireframe screens as artifact for S17 generation.
-   * SD-LEO-FIX-FIX-STAGE-VENTURE-001: Wire postStage15Hook into pipeline.
-   */
-  async _postStageHook_S15_StitchProvision(ventureId) {
-    // SD-LEO-ORCH-REPLACE-GOOGLE-STITCH-001-A: Store wireframe screen data
-    // as venture_artifact instead of calling Stitch APIs. The archetype
-    // generator reads this artifact directly for S17 design generation.
-    const { writeArtifact } = await import('./artifact-persistence-service.js');
-
-    // SD-LEO-INFRA-S15-WIREFRAME-SCREENS-REGRESSION-001 (FR-1): the S15 producer now emits
-    // wireframe_screens in its typed batch (before the 15->16 boundary). This post-hook is now an
-    // IDEMPOTENT FALLBACK for legacy/orchestrator-direct ventures — if a current wireframe_screens
-    // already exists, skip (do not re-write / churn the row).
-    const { data: existingScreens } = await this._supabase
-      .from('venture_artifacts')
-      .select('id')
-      .eq('venture_id', ventureId)
-      .eq('lifecycle_stage', 15)
-      .eq('artifact_type', 'wireframe_screens')
-      .eq('is_current', true)
-      .maybeSingle();
-    if (existingScreens) {
-      this._logger.log('[Worker] S15 post-hook: wireframe_screens already current (producer-owned) — skipping fallback write');
-      return;
-    }
-
-    // Read S15 data from venture_stage_work, falling back to venture_artifacts
-    let s15Work = null;
-    const { data: vsw } = await this._supabase
-      .from('venture_stage_work')
-      .select('advisory_data')
-      .eq('venture_id', ventureId)
-      .eq('lifecycle_stage', 15)
-      .order('created_at', { ascending: false })
-      .limit(1)
-      .maybeSingle();
-
-    if (vsw?.advisory_data && (vsw.advisory_data.wireframes || vsw.advisory_data.screens || vsw.advisory_data.ia_sitemap)) {
-      s15Work = vsw;
-    } else {
-      // Fallback: read from blueprint_wireframes artifact
-      const { data: artifact } = await this._supabase
-        .from('venture_artifacts')
-        .select('artifact_data')
-        .eq('venture_id', ventureId)
-        .eq('lifecycle_stage', 15)
-        .eq('artifact_type', 'blueprint_wireframes')
-        .eq('is_current', true)
-        .maybeSingle();
-      if (artifact?.artifact_data) {
-        s15Work = { advisory_data: artifact.artifact_data };
-      }
-    }
-
-    if (!s15Work?.advisory_data) {
-      this._logger.warn('[Worker] S15 post-hook: no wireframe data found, skipping artifact creation');
-      return;
-    }
-
-    const advisoryData = s15Work.advisory_data;
-
-    // SD-LEO-INFRA-S15-WIREFRAME-SCREENS-REGRESSION-001 (FR-1): shared normalizer — identical screen
-    // shape as the producer's typed-batch write, so the fallback can never drift from the primary path.
-    const { buildWireframeScreensPayload } = await import('./stage-templates/stage-15-screens.js');
-    const screensPayload = buildWireframeScreensPayload(advisoryData);
-    const { screens } = screensPayload;
-
-    // Write wireframe_screens artifact for archetype-generator consumption
-    await writeArtifact(this._supabase, {
-      ventureId,
-      lifecycleStage: 15,
-      artifactType: 'wireframe_screens',
-      title: `Wireframe Screens (${screens.length} screens)`,
-      content: JSON.stringify({ screens, source: 'S15_advisory_data' }),
-      artifactData: screensPayload,
-      qualityScore: 80,
-      validationStatus: 'validated',
-      source: 'stage-15-post-hook',
-      metadata: {
-        screenCount: screens.length,
-        deviceTypes: [...new Set(screens.map(s => s.deviceType))],
-      },
-    });
-
-    this._logger.log(`[Worker] S15 post-hook (fallback): stored ${screens.length} wireframe screens as artifact (Stitch bypassed)`);
-  }
-
-  async _postStageHook_S17_DocGen(ventureId) {
-    const { data: ventureRow } = await this._supabase
-      .from('ventures').select('name').eq('id', ventureId).single();
-    const { generateDocs } = await import('./stage-templates/analysis-steps/stage-17-doc-generation.js');
-    await generateDocs({
-      ventureId,
-      ventureName: ventureRow?.name,
-      supabase: this._supabase,
-      logger: this._logger,
-    });
-    this._logger.log('[Worker] S17 post-stage hook: vision + architecture docs generated');
-
-    // SD-S17-WORKER-STRATEGY-GATE-ORCH-001-A: Strategy gate — block until chairman selects strategy.
-    // Required regardless of the design path: the GVOS composer renders S17 from
-    // venture_gvos_profile after strategy approval.
-    // SD-LEO-REFAC-EXTRACT-S17-ARCHETYPE-001: legacy archetype-generator invocation removed
-    // (GVOS composer is the live path; s17_use_gvos_composer is globally enabled).
-    await this._ensureS17StrategySelected(ventureId);
-  }
-
-  /**
-   * SD-LEO-INFRA-RELIABLE-S19-BUILD-001 / FR-1: auto-draft a draft_seed L2 vision.
-   *
-   * Runs AFTER _postStageHook_S17_DocGen. Only fires for leo_bridge ventures (the
-   * seeded_repo path does NOT use the LEO-SD bridge, so it never needs an L2 vision).
-   * seedDraftL2Vision is idempotent (no-op if any L2 doc already exists) and writes
-   * status='draft_seed' / chairman_approved=false — never active/approved. Non-throwing.
-   *
-   * @param {string} ventureId
-   */
-  async _postStageHook_S17_SeedDraftVision(ventureId) {
-    const logger = this._logger;
-    try {
-      // Resolve build_model via the SINGLE arbiter; skip unless leo_bridge.
-      const { data: s19Work } = await this._supabase
-        .from('venture_stage_work')
-        .select('advisory_data')
-        .eq('venture_id', ventureId)
-        .eq('lifecycle_stage', 19)
-        .maybeSingle();
-      const { data: ventureRow } = await this._supabase
-        .from('ventures').select('name, build_model').eq('id', ventureId).maybeSingle();
-      const { resolveBuildModel } = await import('./bridge/resolve-build-model.js');
-      const buildModel = resolveBuildModel({
-        ventureBuildModel: ventureRow?.build_model,
-        legacyBuildMethod: s19Work?.advisory_data?.build_method,
-      });
-      if (buildModel !== 'leo_bridge') {
-        logger.log(`[Worker] S17 SeedDraftVision: build_model=${buildModel} — skipping (only leo_bridge ventures need an L2 vision for the bridge).`);
-        return;
-      }
-
-      const { seedDraftL2Vision } = await import('./stage-templates/analysis-steps/stage-17-doc-generation.js');
-      await seedDraftL2Vision({
-        supabase: this._supabase,
-        ventureId,
-        ventureName: ventureRow?.name,
-        logger,
-      });
-    } catch (err) {
-      logger.warn(`[Worker] S17 SeedDraftVision hook failed (non-fatal): ${err.message}`);
-    }
-  }
-
-  /**
-   * S17 post-stage hook: freeze the GVOS design snapshot (GVOS FR-6).
-   *
-   * SD-LEO-REFAC-RETIRE-LEGACY-STAGE-001
-   *
-   * The frozen visual identity (archetype + tokens + substrate/accent/typography_voice)
-   * was chosen at S11 (venture_gvos_profile). This writes the immutable
-   * locked_prompt_snapshot + locked_at + locked_version at the S17 transition, the
-   * point FR-6 specifies. Idempotent (skips if already locked) and non-throwing
-   * (mirrors the S11 GvosProfile hook contract) so it never blocks stage advancement.
-   *
-   * Backend port of EHG/src/lib/gvos/snapshot-locker.ts (which is browser-client and was
-   * never invoked → locked_at NULL for all ventures, incl. RPC-auto-advanced ones).
-   * Compliance-category tokens are excluded from the lock (live propagation) by
-   * buildLockedSnapshot.
-   */
-  async _postStageHook_S17_GvosLock(ventureId) {
-    const logger = this._logger;
-    try {
-      const { data: profile, error: profErr } = await this._supabase
-        .from('venture_gvos_profile')
-        .select('venture_id, archetype_id, token_overrides, locked_version, locked_at')
-        .eq('venture_id', ventureId)
-        .maybeSingle();
-
-      if (profErr && profErr.code !== 'PGRST116') {
-        logger.warn('[S17-GvosLock] profile load failed:', profErr.message);
-        return;
-      }
-      if (!profile) {
-        logger.warn(`[S17-GvosLock] no venture_gvos_profile for ${ventureId} — cannot lock (S17 renders the no-profile empty-state).`);
-        return;
-      }
-      if (profile.locked_at) {
-        logger.log(`[S17-GvosLock] Skipping — already locked for venture ${ventureId} (locked_version=${profile.locked_version}).`);
-        return;
-      }
-      if (!profile.archetype_id) {
-        logger.warn(`[S17-GvosLock] venture ${ventureId} has no archetype_id — cannot lock.`);
-        return;
-      }
-
-      const { data: archetype, error: archErr } = await this._supabase
-        .from('gvos_archetypes')
-        .select('id, prompt_token, tokens_required, substrate, accent, typography_voice')
-        .eq('id', profile.archetype_id)
-        .maybeSingle();
-      if (archErr || !archetype) {
-        logger.warn(`[S17-GvosLock] gvos_archetypes load failed for ${profile.archetype_id}: ${archErr?.message || 'not found'}`);
-        return;
-      }
-
-      const tokenNames = Array.isArray(archetype.tokens_required) ? archetype.tokens_required : [];
-      let liveTokens = [];
-      if (tokenNames.length > 0) {
-        const { data: toks, error: tokErr } = await this._supabase
-          .from('gvos_tokens')
-          .select('name, category, version_major, version_minor, version_patch')
-          .in('name', tokenNames);
-        if (tokErr) {
-          logger.warn('[S17-GvosLock] gvos_tokens load failed:', tokErr.message);
-          return;
-        }
-        liveTokens = toks || [];
-      }
-
-      const { buildLockedSnapshot } = await import('../../lib/gvos/snapshot-locker.js');
-      const { locked, excluded_compliance_tokens } = buildLockedSnapshot(
-        archetype,
-        liveTokens,
-        profile.token_overrides ?? {},
-      );
-
-      const newVersion = (profile.locked_version ?? 0) + 1;
-      const lockedAt = new Date().toISOString();
-      const { error: writeErr } = await this._supabase
-        .from('venture_gvos_profile')
-        .update({ locked_prompt_snapshot: locked, locked_at: lockedAt, locked_version: newVersion })
-        .eq('venture_id', ventureId);
-
-      if (writeErr) {
-        // Surface loudly — a swallowed lock failure would leave FR-6 silently unsatisfied.
-        logger.error('[S17-GvosLock] lock write FAILED:', writeErr.message);
-        return;
-      }
-
-      logger.log('[S17-GvosLock] design snapshot locked', {
-        ventureId,
-        locked_version: newVersion,
-        archetype: locked.archetype_prompt_token,
-        tokens: locked.tokens_required.length,
-        excluded_compliance: excluded_compliance_tokens.length,
-      });
-    } catch (err) {
-      logger.warn(`[S17-GvosLock] lock hook failed (non-fatal): ${err.message}`);
-    }
-  }
 
   /**
    * SD-S17-WORKER-STRATEGY-GATE-ORCH-001-A: Ensure strategy selection is complete.
